rename parsing module
