+++ app/presentation_parser/text_comorator.py
